fixed name Clash and one level of wrapping
