+++ spinn_front_end_common/interface/abstract_spinnaker_base.py
@@ -1982,7 +1982,7 @@ class AbstractSpinnakerBase(ConfigHandler, SimulatorInterface):
             inputs["DatabaseFilePath"] = (
                 self._last_run_outputs["DatabaseFilePath"])
         if not self._use_virtual_board:
-            algorithms.append("NotificationProtocol")
+            algorithms.append("CreateNotificationProtocol")
 
         outputs = [
             "NoSyncChanges"
